add mock and cleanup various sdk deps
