+++ store/snapshots/helpers_test.go
@@ -18,8 +18,7 @@ import (
 	sdkerrors "cosmossdk.io/errors"
 	"github.com/cosmos/cosmos-sdk/store/snapshots"
 	snapshottypes "github.com/cosmos/cosmos-sdk/store/snapshots/types"
-	"github.com/cosmos/cosmos-sdk/testutil"
-	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/store/types"
 )
 
 func checksums(slice [][]byte) [][]byte {
@@ -174,7 +173,7 @@ func (m *mockSnapshotter) SetSnapshotInterval(snapshotInterval uint64) {
 // setupBusyManager creates a manager with an empty store that is busy creating a snapshot at height 1.
 // The snapshot will complete when the returned closer is called.
 func setupBusyManager(t *testing.T) *snapshots.Manager {
-	store, err := snapshots.NewStore(db.NewMemDB(), testutil.GetTempDir(t))
+	store, err := snapshots.NewStore(db.NewMemDB(), t.TempDir())
 	require.NoError(t, err)
 	hung := newHungSnapshotter()
 	hung.SetSnapshotInterval(opts.Interval)
@@ -258,7 +257,7 @@ func (s *extSnapshotter) SupportedFormats() []uint32 {
 
 func (s *extSnapshotter) SnapshotExtension(height uint64, payloadWriter snapshottypes.ExtensionPayloadWriter) error {
 	for _, i := range s.state {
-		if err := payloadWriter(sdk.Uint64ToBigEndian(uint64(i))); err != nil {
+		if err := payloadWriter(types.Uint64ToBigEndian(uint64(i))); err != nil {
 			return err
 		}
 	}
@@ -273,7 +272,7 @@ func (s *extSnapshotter) RestoreExtension(height uint64, format uint32, payloadR
 		} else if err != nil {
 			return err
 		}
-		s.state = append(s.state, sdk.BigEndianToUint64(payload))
+		s.state = append(s.state, types.BigEndianToUint64(payload))
 	}
 	// finalize restoration
 	return nil
